pyproject.toml
[tool.poetry]
name = "langchain-lotr-example"
version = "0.1.0"
description = ""
authors = ["Trayan Azarov <trayan.azarov@amikos.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
chromadb = "^0.4.23"
langchain = "^0.1.9"
openai = "^1.12.0"
langchain-openai = "^0.0.8"
sentence-transformers = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
